#15 FOTFIX / 스크립트 타입 enum 클래스화에 따른 변경

+++ Project/Scripts/CBackgroundScript.cpp
@@ -12,7 +12,7 @@ void CBackgroundScript::LoadFromFile(FILE* _File)
 }
 
 CBackgroundScript::CBackgroundScript()
-	: CScript(SCRIPT_TYPE::BACKGROUNDSCRIPT)
+	: CScript((UINT)SCRIPT_TYPE::BACKGROUNDSCRIPT)
 {
 }
